requirements.txt
# Web Framework
flask>=3.0.0
flask-cors>=4.0.0

# Graph Theory
networkx>=3.2

# Scientific Computing & Probability
scipy>=1.11.0
numpy>=1.26.0

# Scryfall API
scrython>=2.0.0
requests>=2.31.0

# Desktop App Packaging
pywebview>=5.0.0

# Development
pytest>=7.4.0
